feat(match): add Phase Blueprint to pre-match; extract shared PhaseBlueprintPanel
Pre-match now has the same tactical depth as the tactics board: the right panel
of the Your Team tab carries the Phase Blueprint (with-ball / without-ball /
transitions settings), editing the same team.tactics_phase via set_tactics_phase.

- Extracts the blueprint UI (fields + PhaseButtonGroup + 3 sections) from
  TacticsRightPanel into a shared PhaseBlueprintPanel; both screens now render it.
- Pre-match seeds local phase state from the team's stored tactics_phase and
  persists edits through the service.

NOTE: the match engine does not yet consume tactics_phase (true on the tactics
screen too) — this is UI parity. Engine consumption is the next task.
